Choose starting spell code moved to TextAdventure.py
I moved the choose a starting spell code to the main script (in TextAdventure.py). I figured it might be better to have it in that one place rather than in each of the character type classes. To facilitate this, I organized the starting spells in a 2D array in the Game class.