added a new file for items to be used in game and traps that the player can come across on adventures
I added a new module for items that can be used in game like potions and any other consumable.  I also added a module for traps that can pop up as the player is trying to progress through the adventure like falling in holes, or getting caught in a bear trap.