applied armor and mag resist to each race and added damage reduction based on armor or mag resist
added a parameter to the character race class to equip starting armor for each character.  Also added to the hit_by function in the character info module that reduces the damage a player takes based on his/her armor or mag resist, and the type of damage that was taken.