bug fixes to the battle function and search function and switched inputs and print statements to run on the server.
Fixed an issue that was keeping the battle class from running correctly.  You can now battle in multiplayer mode without an issue.  I also fixed an issue with appending items you find to your list of items when on an adventure.  I also changed all input and print statements to run on the server so you don't have to input everything twice.