Fixed an issue with printing the character race and character type when the game starts, and a minor spelling error.
Added a charactertype variable to the constructor in the player class so we can print the character type to the other players when the game starts.  Also added the player race to the print statement to allow other players see what race and type you chose at the start of the game.  Also fixed a minor spelling error from a previous commit.