fixed mana and health functionality and the functionality of health and mana potions
if a players health goes above the max health value it is automatically set to the default max health value and if mana goes above the max mana value then it is set back to the default max mana value.  Similiarly if mana or health goes below zero the max values are set to 0.