added a Human race, a chainmail armor set, and made the human race available in game.
Created a new human class that will have a perfect balance of speed endurance and hp as opposed to one primary stat.  Also added chainmail armor which sits between leather and plate armor as far as toughness for the human to wear.  Also made the human available in the Game module as a race the player can choose from in the game.