Added functionality to allow the player to get experience and grow levels.
Added a way for the player to gain experience after killing an opponent.  When the experience hits the max xp the player grows 1 level and the max hp increased, as well as the players max mana.  Also, for every level you gain, it adds 150 extra xp that you will need to grow to the next level (lvl 1: 150, lvl 2: 300, etc)