If players mana is below the mana cost of the attack or spell he/she is trying to use, the user will be prompted that he/she does not have enough mana for that attack
If you mana is below the mana cost of the spell you are trying to use, you will be prompted that you do not have enough mana for that attack.  You can still use basic attacks as they do not use mana.  Once you use a mana potion and increase your mana, the spells will be enabled.