Tweaks to player choosing an attack
Made it so that on_attacking status effect trigger is only activated when a valid attack is chosen. Also made it so that the player can choose another attack if they choose one that is disabled or costs too much mana. Also now skipping a player's turn if they do not have any items or any attacks they can actually use.