Upgrade attack damage upon leveling up, also defined a function to upgrade attacks to allow us to manually upgrade attacks based on the need for it.
Defined an upgrade funciton that upgrades the characters spells upon leveling up.  I also defined an upgrade function to override the base class upgrade funciton in the block class so the block class can upgrade shield amount instead of attack.  This will also work if we want to add heals into the game as well so we can upgrade each spell or attack individually.