added a clues module to find clues after battles or in camps.  Added a function to find clues.  and increased damage in basic attacks module to make battles more difficult.
I made a clues module that will hold pieces of riddles that the player will need to solve at certain points in the game.  The answer to the riddles will be the key to move on.  I also added a function in the adventures1 module to search for clues in camps or get them from defeating enemies.  I also increased the attacks of each monster so they will be able to deal damage to the player and make battles a little more challenging.