Made attack targeting code more concise and versatile
Instead of defining a target type, each attack now has a filter_targets function which given an attack and list of potential targets, filters the list so that only viable targets for that attack are left. Common filters are defined in AttacksInfo.TargetFilters. Each attack targets enemies by default, but the filter_targets function can be overridden in subclasses for different behavior as needed.