Trap avoidance and status effects
If a player its a trap they are prompted to enter a word quickly, if the word is entered correctly they avoid the trap.  If the word is entered incorrectly, the player is hit by the trap, takes damage, and status effect is applied if there is one.  Also added a Design document to keep track of tasks that need to be completed.