Fixed a bug whereby the player would not be able to recieve a clue
Fixed a bug whereby the program would throw an OutOfBounds error when awarding a clue to a player. This was due to the fact that the traps would remove players from the players list as they avoided that trap. The trap threads now utilize a copy of the players list instead to keep track of their activity so the origial players list remains in tact.