Added Bridge Event to Forest level
Added a Bridge Event whereby the party comes upon an old rope bridge spanning the length of a gorge. If more than one player tries to cross at the same time, the bridge collapses and the space becomes unpassable. Otherwise, the players continue across to the next space. The argurments passed to the bridge_event method determine what directions the players move to when they cross or fail to cross respectively.