Updates to Bridge Event: player actions after crossing, player death
Made it so that the player can now take some actions after they have crossed the bridge (such as saying something to their party) and made it so some actions have a different result if the player is already across than if they weren't (for example, you cannot cross the bridge again to go back). Also made it so that if the player attempts to jump across the gorge or something similar, they fall to their death!